+++ test/foundry/newStrategies/FloorFromChainlinkPremiumFixedAmountOrders.t.sol
@@ -34,7 +34,7 @@ contract FloorFromChainlinkPremiumFixedAmountOrdersTest is FloorFromChainlinkPre
         assertTrue(isValid);
         assertEq(errorSelector, bytes4(0));
 
-        vm.expectRevert(abi.encodeWithSelector(IExecutionManager.StrategyNotAvailable.selector, uint16(1)));
+        vm.expectRevert(abi.encodeWithSelector(IExecutionManager.StrategyNotAvailable.selector, 1));
         _executeTakerBid(takerBid, makerAsk, signature);
     }
